+++ StevenDimDoors/mod_pocketDimClient/RenderTransTrapdoor.java
@@ -56,13 +56,14 @@ public class RenderTransTrapdoor extends TileEntitySpecialRenderer
         for (int count = 0; count < 16; ++count)
         {
             GL11.glPushMatrix();
-            float var15 = (float)(16 - count);
+            float var15 = 16 - count;
             float var16 = 0.2625F;
             float var17 = 1.0F / (var15 + 1.0F);
 
             if (count == 0)
             {
                 this.bindTextureByName("/RIFT.png");
+                // move files into assets/modid and change to new ResourceLocation(modid:/RIFT.png)
                 var17 = 0.1F;
                 var15 = 25.0F;
                 var16 = 0.125F;
@@ -73,15 +74,16 @@ public class RenderTransTrapdoor extends TileEntitySpecialRenderer
             if (count == 1)
             {
                 this.bindTextureByName("/WARP.png");
+                // move files into assets/modid and change to new ResourceLocation(modid:/WARP.png)
                 GL11.glEnable(GL11.GL_BLEND);
                 GL11.glBlendFunc(GL11.GL_ONE, GL11.GL_ONE);
                 var16 = .5F;
             }
 
-            GL11.glTranslatef( (float)(Minecraft.getSystemTime() % 200000L) / 200000.0F,0, 0.0F);
-            GL11.glTranslatef(0, (float)(Minecraft.getSystemTime() % 200000L) / 200000.0F, 0.0F);
+            GL11.glTranslatef( Minecraft.getSystemTime() % 200000L / 200000.0F,0, 0.0F);
+            GL11.glTranslatef(0, Minecraft.getSystemTime() % 200000L / 200000.0F, 0.0F);
 
-            GL11.glTranslatef(0,0, (float)(Minecraft.getSystemTime() % 200000L) / 200000.0F);
+            GL11.glTranslatef(0,0, Minecraft.getSystemTime() % 200000L / 200000.0F);
 
             GL11.glTexGeni(GL11.GL_S, GL11.GL_TEXTURE_GEN_MODE, GL11.GL_OBJECT_LINEAR);
             GL11.glTexGeni(GL11.GL_T, GL11.GL_TEXTURE_GEN_MODE, GL11.GL_OBJECT_LINEAR);
@@ -102,10 +104,10 @@ public class RenderTransTrapdoor extends TileEntitySpecialRenderer
             GL11.glMatrixMode(GL11.GL_TEXTURE);
             GL11.glPushMatrix();
             GL11.glLoadIdentity();
-            GL11.glTranslatef(0.0F, (float)(Minecraft.getSystemTime() % 200000L) / 200000.0F*var15, 0.0F);
+            GL11.glTranslatef(0.0F, Minecraft.getSystemTime() % 200000L / 200000.0F*var15, 0.0F);
             GL11.glScalef(var16, var16, var16);
             GL11.glTranslatef(0.5F, 0.5F, 0.5F);
-            GL11.glRotatef((float)(count * count * 4321 + count * 9) * 2.0F, 0.0F, 0.0F, 1.0F);
+            GL11.glRotatef((count * count * 4321 + count * 9) * 2.0F, 0.0F, 0.0F, 1.0F);
             GL11.glTranslatef(0.5F, 0.5F, 0.5F);
         
             GL11.glBegin(GL11.GL_QUADS);
